Rename search condition modal component
